Start with ancestorSets of each transaction
Instead of starting the search only with the ancestorless transactions,
we start with the ancestorSet of each transaction as a starting point.
This allows us to immediately prune a lot of useless descendants and
find outstanding CPFP pairs.

Also caches weight, fees, and feerates of candidateSets so it is only
calculated once.

Reduces runtime by 2/3.